core/pyproject.toml
[tool.poetry]
name = "aws-ddk-core"
version = "0.6.2"
description = "AWS DataOps Development Kit - Core"
authors = [
    "AWS Professional Services <aws-proserve-opensource@amazon.com>"
]
license = "Apache-2.0"
readme = "README.md"

include = ["README.md", "LICENSE.txt", "NOTICE.txt", "THIRD_PARTY.txt"]

exclude = ["*.so", "*.pyc", "*~", "#*", ".git*", ".coverage*", "DS_Store", "__pycache__"]

keywords = ["aws", "ddk", "dataops", "core"]

homepage = "https://github.com/awslabs/aws-ddk"
repository = "https://github.com/awslabs/aws-ddk"
documentation = "https://awslabs.github.io/aws-ddk/"

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Utilities",
    "Typing :: Typed",
]

[tool.poetry.dependencies]
python = ">=3.7.1, <3.12"
boto3 = "^1.24.11"
boto3-stubs-lite = {version = "*", extras = ["lambda"]}
botocore = "^1.27.11"
aws-cdk-lib = "^2.54.0"
"aws-cdk.aws-glue-alpha" = "^2.54.0a0"
"aws-cdk.aws-kinesisfirehose-alpha" = "^2.54.0a0"
"aws-cdk.aws-kinesisfirehose-destinations-alpha" = "^2.54.0a0"
marshmallow = "^3.14.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
